+++ wemake_python_styleguide/violations/naming.py
@@ -120,59 +120,6 @@ Pattern matching
 
 .. currentmodule:: wemake_python_styleguide.violations.naming
 
-Summary
--------
-
-.. autosummary::
-   :nosignatures:
-
-   WrongModuleNameViolation
-   WrongModuleMagicNameViolation
-   WrongModuleNamePatternViolation
-   WrongVariableNameViolation
-   TooShortNameViolation
-   PrivateNameViolation
-   SameAliasImportViolation
-   UnderscoredNumberNameViolation
-   UpperCaseAttributeViolation
-   ConsecutiveUnderscoresInNameViolation
-   ReservedArgumentNameViolation
-   TooLongNameViolation
-   UnicodeNameViolation
-   TrailingUnderscoreViolation
-   UnusedVariableIsUsedViolation
-   UnusedVariableIsDefinedViolation
-   WrongUnusedVariableNameViolation
-   UnreadableNameViolation
-   BuiltinShadowingViolation
-
-Module names
-------------
-
-.. autoclass:: WrongModuleNameViolation
-.. autoclass:: WrongModuleMagicNameViolation
-.. autoclass:: WrongModuleNamePatternViolation
-
-General names
--------------
-
-.. autoclass:: WrongVariableNameViolation
-.. autoclass:: TooShortNameViolation
-.. autoclass:: PrivateNameViolation
-.. autoclass:: SameAliasImportViolation
-.. autoclass:: UnderscoredNumberNameViolation
-.. autoclass:: UpperCaseAttributeViolation
-.. autoclass:: ConsecutiveUnderscoresInNameViolation
-.. autoclass:: ReservedArgumentNameViolation
-.. autoclass:: TooLongNameViolation
-.. autoclass:: UnicodeNameViolation
-.. autoclass:: TrailingUnderscoreViolation
-.. autoclass:: UnusedVariableIsUsedViolation
-.. autoclass:: UnusedVariableIsDefinedViolation
-.. autoclass:: WrongUnusedVariableNameViolation
-.. autoclass:: UnreadableNameViolation
-.. autoclass:: BuiltinShadowingViolation
-
 """
 
 from typing import final