+++ wemake_python_styleguide/violations/oop.py
@@ -6,53 +6,6 @@ and using objects correctly. That's the place we collect these kind of rules.
 
 .. currentmodule:: wemake_python_styleguide.violations.oop
 
-Summary
--------
-
-.. autosummary::
-   :nosignatures:
-
-   BuiltinSubclassViolation
-   ShadowedClassAttributeViolation
-   StaticMethodViolation
-   BadMagicMethodViolation
-   WrongClassBodyContentViolation
-   MethodWithoutArgumentsViolation
-   WrongBaseClassViolation
-   WrongSlotsViolation
-   WrongSuperCallViolation
-   DirectMagicAttributeAccessViolation
-   AsyncMagicMethodViolation
-   YieldMagicMethodViolation
-   UselessOverwrittenMethodViolation
-   WrongSuperCallAccessViolation
-   WrongDescriptorDecoratorViolation
-   UnpythonicGetterSetterViolation
-   BuggySuperContextViolation
-   LambdaAttributeAssignedViolation
-
-Respect your objects
---------------------
-
-.. autoclass:: BuiltinSubclassViolation
-.. autoclass:: ShadowedClassAttributeViolation
-.. autoclass:: StaticMethodViolation
-.. autoclass:: BadMagicMethodViolation
-.. autoclass:: WrongClassBodyContentViolation
-.. autoclass:: MethodWithoutArgumentsViolation
-.. autoclass:: WrongBaseClassViolation
-.. autoclass:: WrongSlotsViolation
-.. autoclass:: WrongSuperCallViolation
-.. autoclass:: DirectMagicAttributeAccessViolation
-.. autoclass:: AsyncMagicMethodViolation
-.. autoclass:: YieldMagicMethodViolation
-.. autoclass:: UselessOverwrittenMethodViolation
-.. autoclass:: WrongSuperCallAccessViolation
-.. autoclass:: WrongDescriptorDecoratorViolation
-.. autoclass:: UnpythonicGetterSetterViolation
-.. autoclass:: BuggySuperContextViolation
-.. autoclass:: LambdaAttributeAssignedViolation
-
 """
 
 from typing import final