+++ wemake_python_styleguide/visitors/ast/builtins.py
@@ -61,6 +61,7 @@ class WrongStringVisitor(base.BaseNodeVisitor):
 
     def visit_any_string(self, node: ast.Constant) -> None:
         """Forbids incorrect usage of strings."""
+        assert isinstance(node.value, str | bytes)  # for mypy  # noqa: S101
         text_data = source.render_string(node.value)
         self._check_is_alphabet(node, text_data)
         self.generic_visit(node)
@@ -208,7 +209,9 @@ class WrongNumberVisitor(base.BaseNodeTokenVisitor):
             or node.value in constants.MAGIC_NUMBERS_WHITELIST
             or is_non_magic
             or check_is_node_in_specific_annotation(
-                parent, 'Literal', self._allowed_modules_to_literal_type_hint
+                parent,
+                'Literal',
+                self._allowed_modules_to_literal_type_hint,
             )
         ):
             return