+++ CHANGELOG.md
@@ -16,9 +16,9 @@ Semantic versioning in our case means:
   But, in the future we might change the configuration names/logic,
   change the client facing API, change code conventions significantly, etc.
 
-## 1.3.0 - WIP
+## WIP
 
-### Features
+### Misc
 
 - Adds custom Sphinx directive `.. plugincodes` for violation rendering, #1318