+++ CHANGELOG.md
@@ -16,7 +16,12 @@ Semantic versioning in our case means:
   But, in the future we might change the configuration names/logic,
   change the client facing API, change code conventions significantly, etc.
 
+
 ## WIP
+
+### Features
+
+- Default value of `MAX_CONDITIONS` for `WPS222` was increased from 4 to 5
 
 ### Misc