+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Semantic versioning in our case means:
   But, in the future we might change the configuration names/logic,
   change the client facing API, change code conventions significantly, etc.
 
+
 ## WIP
 
 ### Misc